fix Album

+++ src/albums/albums.controller.ts
@@ -1,34 +1,66 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, UseInterceptors, ClassSerializerInterceptor } from '@nestjs/common';
+import { NotFoundException } from '@nestjs/common/exceptions';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { UsersService } from 'src/users/users.service';
 import { AlbumsService } from './albums.service';
 import { CreateAlbumDto } from './dto/create-album.dto';
 import { UpdateAlbumDto } from './dto/update-album.dto';
 
+  /** Création d'un nouvel album nécessite :
+   * * 
+   */
+  @UseGuards(JwtAuthGuard) // doit être connecté/enregistré
 @Controller('albums')
 export class AlbumsController {
-  constructor(private readonly albumsService: AlbumsService) {}
+  constructor(
+    private readonly albumsService: AlbumsService,
+    private readonly userService: UsersService) {}
+
 
   @Post()
-  create(@Body() createAlbumDto: CreateAlbumDto) {
+  async create(@Body() createAlbumDto: CreateAlbumDto) {
     return this.albumsService.create(createAlbumDto);
   }
 
   @Get()
-  findAll() {
-    return this.albumsService.findAll();
+  async findAll() {
+    const allAlbums = await this.albumsService.findAll();
+    if (!allAlbums){
+      throw new NotFoundException(`Albums not found.`);
+    }
+    return allAlbums;
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.albumsService.findOne(+id);
+  async findOne(@Param('id') id: string) {
+    const oneAlbum = await this.albumsService.findOne(+id);
+    if(!oneAlbum){
+      throw new NotFoundException(`Album with id ${id} not found.`);
+    }
+    return oneAlbum;
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateAlbumDto: UpdateAlbumDto) {
-    return this.albumsService.update(+id, updateAlbumDto);
+  async update(@Param('id') id: string, @Body() updateAlbumDto: UpdateAlbumDto) {
+    const albumExist = await this.findOne(id);
+    if (!albumExist){
+      throw new NotFoundException(`Album with id ${id} not found.`);
+    }
+    const updateAlbum = await this.albumsService.update(+id, updateAlbumDto);
+    return updateAlbum;
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.albumsService.remove(+id);
+  async remove(@Param('id') id: string) {
+    const albumExist = await this.findOne(id);
+    if (!albumExist){
+      throw new NotFoundException(`Album with id ${id} not found.`);
+    }
+    const deletedAlbum = await this.albumsService.remove(+id);
+    return {
+      statusCode: 200,
+      message: `deleted album`,
+      data: deletedAlbum
+    }
   }
 }
